Removed DefaultDiffAdapter. Added autoScrollToTop into DiffAdapter
